Windows: add support for a Windows-wide configuration
Between the libgit2 and the Git for Windows project, there has been a
discussion how we could share Git configuration to avoid duplication (or
worse: skew).

Earlier, libgit2 was nice enough to just re-use Git for Windows'

	C:\Program Files (x86)\Git\etc\gitconfig

but with the big changes that went into Git for Windows 2.x, there would
have been more paths to search, as there now was a 64-bit version in
addition to a 32-bit one, and the corresponding config files are
`%PROGRAMFILES%\Git\mingw64\etc` and `...\mingw32\etc`, respectively.

Worse: there are portable Git for Windows versions out there which live
in totally unrelated directories, still.

Therefore we came to a consensus to use `%PROGRAMDATA%\Git\config` as the
location for shared Git settings that are of wider interest than just Git
for Windows, read: for other Git implementations.

Of course, the configuration in `%PROGRAMDATA%\Git\config` has the
widest reach, therefore it must take the lowest precedence, i.e. Git for
Windows can still override settings in its `etc/gitconfig` file.

Git for Windows supports this since v2.5.0 (i.e. since the first v2.x),
and libgit2 supports it as of v0.28.1:
https://github.com/libgit2/libgit2/commit/8c7c5fa585

Helped-by: Andreas Heiduk <asheiduk@gmail.com>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>