mingw: abolish the previous environment handling code
Let's use the code we developed for MSVC, it is much cleaner.

This still needs a good commit message (stating that we do not call
getenv() all that often anymore), and it needs to be moved far closer to
the branch point, and the msvc branches need to be adjusted.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>