nedmalloc: do not try to override the system allocator
When we integrated nedmalloc, it still supported dlmalloc's feature
where you could still use malloc()/calloc()/realloc()/free() and all of
a sudden would use nedmalloc instead of the system allocator. This
feature was toggled by defining the REPLACE_SYSTEM_ALLOCATOR constant.

However, as of nedmalloc's d8806bc8b3c (Added USE_MAGIC_HEADERS which
allows nedmalloc to handle freeing a system allocated block. [...],
2009-09-15) this feature had been broken, and furthermore, Git would not
be able to use it after this commit, anyway, because it is incompatible
with ENABLE_TOLERANT_NEDMALLOC (which we require to be turned off).

So let's bite the bullet in preparation for upgrading to a newer
nedmalloc version and use nedmalloc(), nedcalloc(), nedrealloc() and
nedfree() instead, when USE_NED_ALLOCATOR is in effect.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>