http: default to use HTTP/1.1, to help NTLM users
In particular on Windows, where NTLM authentication is a very convenient
thing, it is important to keep working with NTLM-authenticated remote
repositories. However, NTLM is incompatible with HTTP/2 (which is now
the default in cURL).

https://docs.microsoft.com/en-us/iis/get-started/whats-new-in-iis-10/http2-on-iis#when-is-http2-not-supported
says:

	Windows authentication (NTLM/Kerberos/Negotiate) is not
	supported with HTTP/2. In this case IIS will fall back to
	HTTP/1.1.

Unfortunately, cURL v7.62.0 does not accommodate that, so for now, let's
just force Git to use HTTP/1.1 (unless the user overrides this via the
`http.version` config setting).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>