t9902: 'send-email' test case requires PERL
With NO_PERL, the `git send-email` script errors out with code 128,
mentioning that Git was built without Perl support.

Therefore, when the completion tries to ask for possible completions via
`git send-email --git-completion-helper`, it won't provide what is
necessary for that test case to pass.

So let's mark it with the PERL prerequisite.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>