archive: factor out writing blocks into a separate function
The `git archive --format=tgz` command spawns `gzip` to perform the
actual compression. However, the MinGit flavor of Git for Windows
comes without `gzip` bundled inside.

To help with that, we will teach `git archive` to let zlib perform the
actual compression.

In preparation for this, we consolidate all the block writes into the
function `write_block_or_die()`.

Note: .tar files have a well-defined, fixed block size. For that reason,
it does not make any sense to pass anything but a fully-populated,
full-length block to the `write_block_or_die()` function, and we can
save ourselves some future trouble by not even allowing to pass an
incorrect `size` parameter to it.

Signed-off-by: Rohit Ashiwal <rohit.ashiwal265@gmail.com>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>