Add support for very simple Inter-Process Communication
For the upcoming `fsmonitor--daemon` command, we need some way to send a
command from some Git process to that daemon, and to receive an answer.
In other words, we need to support Inter-Process Communication.

By nature, Inter-Process Communication requires OS-specific
implementations.

Instead of using full-blown TCP/UDP socket communication (which comes
with all kinds of challenges such as: restricting access permissions,
collisions when multiple instances might want to listen on the same
port, etc), we use named pipes on Windows, and plan on implementing a
backend based on Unix sockets for Linux/Unix later.

For now, we use a super-simple protocol: identified by a path into
.git/, the daemon listens for clients, reads one NUL-terminated command,
then sends the answer and closes the connection.

Note: we do _not_ use the `pkt-line` protocol: while it is almost
libified, there is at least one code path starting from
`packet_read_line_gently()` that dies in `get_packet_data()`, and we
really, really do not want our simple IPC client to take down the entire
Git process with it when it could not talk to `fsmonitor--daemon`, but
would rather instead fall back to a full scan of all tracked files.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>