nedmalloc: help diagnose small buffer overruns
Frequently, when modifying bytes located *just* after a `malloc()`ed
region, it goes undetected because that does not cause problems with GNU
libc's `malloc()` nor with valgrind (valgrind actually does not detect
these bugs, it seems).

However, when using nedmalloc, this type of corruption leads to serious
problems later on, when releasing the allocated memory, because that is
exactly the location where nedmalloc stores metadata about the region,
and if it is corrupted, well, that leads to segmentation faults.

Obviously, these segmentation faults happen _way_ after the actual
corruption takes place.

With this function, it is possible to diagnose such issues, if a
reliable reproducer is available: the `check_nedmalloc_invariants()`
function can be called with the pointer to the allocated region, to see
whether it has been corrupted yet.

This works best when compiled with `FULLSANITYCHECKS` defined in
`compat/nedmalloc/nedmalloc.c`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>