pcre2: allow overriding the system allocator
On Windows, we compile Git using nedmalloc (via `USE_NED_ALLOCATOR`),
but PCRE2 still uses the system allocator. This does not pose any major
problem yet because we never let PCRE2 allocate memory that we then want
to release manually, calling `free()` directly.

This is about to change, though, and in preparation, let's use the
appropriate PCRE2 API facilities to make sure that PCRE2 also uses the
nedmalloc when Git wants to use it.

As there are other ways to override the system allocator than
`USE_NED_ALLOCATOR`, we choose to specify the allocator we want to use
specifically, even if we still use the system allocator.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>