archive (internal gzip): compress in a separate thread
Rather than compressing the file contents in the same thread as
unpacking the objects from Git's object database, we can distribute the
load onto two threads, speeding up the operation a bit in the process.

Note: to allow forcing single-threaded operations, this mode now
respects the `pack.threads` setting.

As to the speed impact, with zlib v1.2.11, the extra thread seems to be
slightly faster than the version without the extra thread, but it seems
to be still about 20% slower than spawning `gzip` in this developer's
setup.

However, as per a recent mail on the zlib mailing list (see
http://madler.net/pipermail/zlib-devel_madler.net/2019-December/003308.html
for details), the situation will (hopefully) improve soon-ish. Patches
provided by Intel at https://github.com/jtkukunas/zlib, for example,
seem to reverse the picture in this developer's setup: even the
single-threaded `:internal-gzip:` is now about 5% faster than the mode
spawning `gzip`.

Initial-patch-by: René Scharfe <l.s.r@web.de>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>