t: protect against use of test_path_is_hidden outside MINGW
This test helper function is meant to test if the "hidden"
atttribute is set (or unset) as expected on Windows, and use of it
outside MINGW prerequisite is an error.  Ensure we have the prereq
and trigger a BUG otherwise.

It is tempting to instead replace its implementation with something
like

	if test_have_prereq MINGW
	then
		... current Windows specific code ...
	else
		# ls without -a/-A hides paths that begin with a dot
		case "$(basename "$1")" in
		.*) return 0 ;;
		esac
	fi
	return 1

but one test in t0001 is designed specifically that with an option
"repository/.git" of a newly created repository is *not* hidden on
Windows, which means that it is impossible to make that feature work
on POSIX systems and the above won't test what we want to test.

Signed-off-by: Junio C Hamano <gitster@pobox.com>