t5515: use the default branch name `main` for testing
Ideally, we would want to be able to verify that every test in Git's
test suite works when overriding the default branch name, in preparation
for the long-term goal to replace that branch name by an inclusive one.

In the particular case of t5515, this is anywhere between tricky and
impossible: the default branch name is not only encoded in some file
names inside `t5515/`, it is also enshrined in the hard-coded
SHA-1/SHA-256 values. Also, some of the output that is compared against
pre-recorded output depends on the alphabetical order of branch names
(one of which is the default branch).

So let's adjust that particular test script to hard-code the default
branch name `main` instead of trying to make it work with arbitrary
default branch names.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>