t3200: avoid variations of the `master` branch name
To avoid branch names with a loaded history, we already started to shift
in the direction of using `main` as the default branch name.

The `t3200-branch.sh` script uses variations of this name for branches
other than the default one. So let's address these variations first, as
they are the "lowest-hanging fruits".

In this particular script, this rename requires a couple of non-trivial
adjustments, though, as the aligned output depends on the maximum length
of the displayed branches (which we now changed), and also on the
alphabetical order (which we now changed, too).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>