tests: avoid variations of the `master` branch name
The term `master` has a loaded history that serves as a constant
reminder of racial injustice. Git wants no part in perpetuating this,
therefore we are preparing to use a different default branch name.

The test suite uses variations of this name for branches other than the
default one. It is very easy to replace those instances because they do
not require any changes outside the test script, so let's do that.

Seeing as `main` seems to be materializing as a good replacement for the
default branch name, we replace the variations of `master` with
variations of `main`.

This trick was performed by this (GNU) sed invocation:

	$ sed -i 's/master\([a-z0-9]\)/main\1/g' t/t*.sh

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>