t/test-terminal: avoid non-inclusive language
Git does not want to perpetuate the harm inflicted by the terms `master`
and `slave`.

In this instance, the sad part is that the term `slave` is enshrined in
IO::Pty's API. We simply cannot avoid using that word here. But at least
we can get rid of the usage of the word `master` and hope that IO::Pty
will be eventually fixed, too.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>