contrib/svn-fe: use configured default branch name
We just introduced a config variable to be able to override the default
branch name. Even if things in `contrib/` are technically not maintained
actively, it is easy to support that, so let's do that.

The only tricky part about this patch is that we need to read the Git
config, which requires all the initializations of common-main.c.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>