tests: prepare aligned mentions of the default branch name
In some tests, the default branch name is part of aligned output. As we
want to change the default branch name to `main`, which is two
characters shorter than the old default branch name, we will have to
adjust those tests.

Since we use the original default branch name until the entire test
suite has bee adjusted accordingly, the touched test cases need to be
guarded by a prereq (that is so far disabled so that they are skipped).

The test cases that depend on the test cases newly guarded by that
prereq naturally have to be guarded, too.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>